+++ packages/proto-signing/types/index.d.ts
@@ -4,5 +4,5 @@ export { EncodeObject, Registry } from "./registry";
 export { DirectSecp256k1HdWallet } from "./directsecp256k1hdwallet";
 export { DirectSecp256k1Wallet } from "./directsecp256k1wallet";
 export { decodePubkey, encodePubkey } from "./pubkey";
-export { isOfflineDirectSigner, OfflineDirectSigner, OfflineSigner } from "./signer";
+export { DirectSignResponse, isOfflineDirectSigner, OfflineDirectSigner, OfflineSigner } from "./signer";
 export { makeAuthInfoBytes, makeSignBytes, makeSignDoc } from "./signing";